Changes from CR + splited door file into multiple files with different types of doors
