+++ src/components/schedule-visualizer/door/door-style.tsx
@@ -1,0 +1,24 @@
+import { makeStyles } from '@material-ui/core';
+
+const doorStyles = makeStyles(() => ({
+  doorMarker: {
+    cursor: 'pointer',
+    pointerEvents: 'auto',
+  },
+  door: {
+    strokeWidth: '0.2',
+  },
+  doorOpen: {
+    stroke: '#AFDDAE',
+    strokeDasharray: 0.1,
+  },
+  doorClose: {
+    stroke: '#BC4812',
+  },
+  doorProcess: {
+    stroke: '#E9CE9F',
+    strokeDasharray: 0.3,
+  },
+}));
+
+export default doorStyles;
