+++ src/components/schedule-visualizer/door/door.test.tsx
@@ -1,6 +1,6 @@
 import { createMount } from '@material-ui/core/test-utils';
-import { Spinner } from './spinner';
-import Door, { SingleHingeDoor, DoorStates } from './door';
+import { Spinner } from '../spinner';
+import Door, { SingleHingeDoor, DoorMode } from './door';
 import React from 'react';
 
 const mainDoor = {
@@ -37,7 +37,7 @@ test('Trigger click event', async () => {
 test('Checks spinner activation when the door its opening or closing', async () => {
   const wrapper = mount(
     <svg>
-      <Door door={mainDoor} currentMode={DoorStates.PROCESS} />
+      <Door door={mainDoor} currentMode={DoorMode.PROCESS} />
     </svg>,
   );
   expect(
@@ -54,7 +54,7 @@ describe('Checks assignation of styles on different states of SingleHingeDoor', 
   test('Style of SingleHingeDoor on Open', () => {
     const wrapper = mount(
       <svg>
-        <SingleHingeDoor door={mainDoor} currentMode={DoorStates.OPEN} />
+        <SingleHingeDoor door={mainDoor} currentMode={DoorMode.OPEN} />
       </svg>,
     );
     const singleDoorSVGLine = wrapper
@@ -68,7 +68,7 @@ describe('Checks assignation of styles on different states of SingleHingeDoor', 
   test('Style of SingleHingeDoor on Close', () => {
     const wrapper = mount(
       <svg>
-        <SingleHingeDoor door={mainDoor} currentMode={DoorStates.CLOSE} />
+        <SingleHingeDoor door={mainDoor} currentMode={DoorMode.CLOSE} />
       </svg>,
     );
     const singleDoorSVGLine = wrapper
@@ -82,7 +82,7 @@ describe('Checks assignation of styles on different states of SingleHingeDoor', 
   test('Style of SingleHingeDoor on Process', () => {
     const wrapper = mount(
       <svg>
-        <SingleHingeDoor door={mainDoor} currentMode={DoorStates.PROCESS} />
+        <SingleHingeDoor door={mainDoor} currentMode={DoorMode.PROCESS} />
       </svg>,
     );
     const singleDoorSVGLine = wrapper
